+++ web-ui/src/components/event-components/elimination-scoring/EliminationScoringEventFrame.tsx
@@ -120,6 +120,7 @@ export const EliminationScoringEventFrame = ({
   });
 
   useEffect(() => {
+    setLoadingNextTick(true);
     getEventRecords({ eventId: event.id, playbackTick: replayTick });
   }, [event.id, replayTick]);