+++ web-ui/src/components/event-components/endurance-race/EnduranceRaceEventFrame.tsx
@@ -101,6 +101,7 @@ export const EnduranceRaceEventFrame: React.FC<EnduranceRaceEventFrameProps> = (
   });
 
   useEffect(() => {
+    setLoadingNextTick(true);
     getEventRecords({ eventId: event.id, playbackTick: replayTick });
   }, [event.id, replayTick]);