+++ web-ui/src/components/event-components/sprint-race/SprintRaceEventFrame.tsx
@@ -101,6 +101,7 @@ export const SprintRaceEventFrame: React.FC<SprintRaceEventFrameProps> = ({
   });
 
   useEffect(() => {
+    setLoadingNextTick(true);
     getEventRecords({ eventId: event.id, playbackTick: replayTick });
   }, [event.id, replayTick]);