+++ CHANGELOG.md
@@ -575,3 +575,4 @@
 - Added keyboard shortcuts to replay and event controls in desktop mode
 - Modified progress button texts to be shorter on mobile
 - Show loading state for event record loading if those records not manually generated
+- Fix max tick estimation for quartered events